Refactor code, but maybe with some ¿errors?
